+++ src/ts/Ads/Views/OpenMeasurement/AdmobOpenMeasurementController.ts
@@ -76,13 +76,12 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
     public injectVerificationResources(verificationResources: IVerificationScriptResource[]) {
         const omVendors: string[] = [];
         verificationResources.forEach((resource) => {
-            const om = new OpenMeasurement<AdMobCampaign>(this._platform, this._core, this._clientInfo, this._campaign, this._placement, this._deviceInfo, this._request, resource.vendorKey);
+            const om = new OpenMeasurement<AdMobCampaign>(this._platform, this._core, this._clientInfo, this._campaign, this._placement, this._deviceInfo, this._thirdPartyEventManager, resource.vendorKey);
             this.setupOMInstance(om, resource);
             if (OpenMeasurementUtilities.isDoubleClickGoogle(resource.vendorKey)) {
-                SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickInstanceCreated,
-                [
-                    SDKMetrics.createAdsSdkTag('dckey', OpenMeasurementUtilities.getDcKeyMetricTag(resource.vendorKey))
-                ]);
+                SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickInstanceCreated, {
+                    'dckey': OpenMeasurementUtilities.getDcKeyMetricTag(resource.vendorKey)
+                });
             }
             omVendors.push(resource.vendorKey);
         });
@@ -174,10 +173,9 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
             this._omInstances.forEach((om) => {
                 const verificationresource = om.getVerificationResource();
                 if (OpenMeasurementUtilities.isDoubleClickGoogle(verificationresource.vendorKey)) {
-                    SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMImpressions,
-                    [
-                        SDKMetrics.createAdsSdkTag('dckey', OpenMeasurementUtilities.getDcKeyMetricTag(verificationresource.vendorKey))
-                    ]);
+                    SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMImpressions, {
+                        'dckey': OpenMeasurementUtilities.getDcKeyMetricTag(verificationresource.vendorKey)
+                    });
                 }
             });
             super.impression(impressionObject);
@@ -192,10 +190,9 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
             this._omInstances.forEach((om) => {
                 const verificationresource = om.getVerificationResource();
                 if (OpenMeasurementUtilities.isDoubleClickGoogle((verificationresource.vendorKey))) {
-                    SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMImpressions,
-                    [
-                        SDKMetrics.createAdsSdkTag('dckey', OpenMeasurementUtilities.getDcKeyMetricTag(verificationresource.vendorKey))
-                    ]);
+                    SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMImpressions, {
+                        'dckey': OpenMeasurementUtilities.getDcKeyMetricTag(verificationresource.vendorKey)
+                    });
                 }
             });
             super.impression(impressionObject);
@@ -217,10 +214,9 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
             om.sessionStart(event);
 
             if (OpenMeasurementUtilities.isDoubleClickGoogle(verificationresource.vendorKey)) {
-                SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMStarts,
-                [
-                    SDKMetrics.createAdsSdkTag('dckey', OpenMeasurementUtilities.getDcKeyMetricTag(verificationresource.vendorKey))
-                ]);
+                SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMStarts, {
+                    'dckey': OpenMeasurementUtilities.getDcKeyMetricTag(verificationresource.vendorKey)
+                });
             }
         });
         SDKMetrics.reportMetricEvent(AdmobMetric.AdmobOMSessionStart);